Modularize gel detection and add docs
